requirements.txt
openai
python-dotenv
pydub
opencv-python
